Switch mesh descriptor to use 16bit floats instead of 32bit floats

+++ Packages/com.stackray.text/Stackray.Text/VertexIndex.cs
@@ -1,11 +1,10 @@
 ﻿using Unity.Entities;
-using Unity.Mathematics;
 
 namespace Stackray.Text {
   [InternalBufferCapacity(54)]    // 54 is the worst case scenario for 9-slice sprite
   public struct VertexIndex : IBufferElementData {
-    public int Value;
-    public static implicit operator VertexIndex(int v) { return new VertexIndex { Value = v }; }
+    public ushort Value;
+    public static implicit operator VertexIndex(int v) { return new VertexIndex { Value = (ushort)v }; }
     public static implicit operator int(VertexIndex v) { return v.Value; }
   }
 }
